+++ LibOS/shim/src/fs/str/fs.c
@@ -87,10 +87,9 @@ int str_close (struct shim_handle * hdl)
     return 0;
 }
 
-int str_read (struct shim_handle * hdl, void * buf,
-              size_t count)
+ssize_t str_read (struct shim_handle * hdl, void * buf, size_t count)
 {
-    int ret = 0;
+    ssize_t ret = 0;
 
     if (!(hdl->acc_mode && MAY_READ)) {
         ret = -EACCES;
@@ -113,8 +112,8 @@ int str_read (struct shim_handle * hdl, void * buf,
     if (!strhdl->ptr)
         strhdl->ptr = data->str;
 
-    int offset = strhdl->ptr - data->str;
-    int remain = data->len - offset;
+    off_t  offset = strhdl->ptr - data->str;
+    size_t remain = data->len - offset;
 
     if (count >= remain) {
         memcpy(buf, strhdl->ptr, remain);
@@ -133,8 +132,7 @@ out:
     return ret;
 }
 
-int str_write (struct shim_handle * hdl, const void * buf,
-               size_t count)
+ssize_t str_write (struct shim_handle * hdl, const void * buf, size_t count)
 {
     if (!(hdl->acc_mode && MAY_WRITE))
         return -EACCES;
@@ -183,8 +181,7 @@ int str_write (struct shim_handle * hdl, const void * buf,
     return count;
 }
 
-int str_seek (struct shim_handle * hdl, off_t offset,
-              int whence)
+off_t str_seek (struct shim_handle * hdl, off_t offset, int whence)
 {
     struct shim_str_handle * strhdl = &hdl->info.str;
